backend/requirements.txt
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
csscompressor==0.9.5
cycler==0.11.0
distlib==0.3.8
docx2txt==0.8
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.2.5
Flask-Cors==4.0.1
fonttools==4.38.0
ginger==0.0.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.3.2
jsmin==3.0.1
kiwisolver==1.4.5
libsass==0.22.0
MarkupSafe==2.1.5
matplotlib==3.5.3
nltk==3.8.1
numpy==1.21.6
openpyxl==3.1.2
packaging==24.0
pandas==1.3.5
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
platformdirs==4.0.0
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.66.4
typing_extensions==4.7.1
virtualenv==20.26.1
Wand==0.6.13
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.15.0